error handling for the key
I put the process of getting the key input into a function and then implemented error handling in the case of improper input. I also added additional characters to the variable that stores all of the potential input options. It will currently not encode the following 4 special characters: \'/".

I also moved the user_input process into a function and made it so that the program will restart once it finishes.